Move LocalFeedLoader (and FeedStore protocol) to its own file in prod
